requirements.txt
flask==2.0.1
google-generativeai==0.3.0
python-docx==0.8.11
openpyxl==3.0.9
reportlab==3.6.8